Отдельный файл для класса с дополнительным заданием.
